chore: optimize database testing
