+++ .github/workflows/ci.yaml
@@ -5,7 +5,14 @@ on:
   push:
     branches:
       - main
+concurrency:
+  group: test-${{ github.head_ref }}
+  cancel-in-progress: true
 
+env:
+  PYTHONUNBUFFERED: "1"
+  FORCE_COLOR: "1"
+  SKIP_DOCKER_COMPOSE: "1"
 jobs:
   validate:
     runs-on: ubuntu-latest
@@ -16,7 +23,7 @@ jobs:
 
       - uses: actions/setup-python@v5
         with:
-          python-version: "3.11"
+          python-version: "3.12"
 
       - uses: pre-commit/action@v3.0.1
 
@@ -26,7 +33,7 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        python-version: ["3.11"]
+        python-version: ["3.12"]
     steps:
       - name: Check out repository
         uses: actions/checkout@v4
@@ -55,8 +62,17 @@ jobs:
       - name: Install dependencies
         run: pdm install -G:all
 
+      - if: runner.os == 'Linux'
+        name: Start Database Services
+        run: sudo docker compose -f tests/docker-compose.yml up -d
+
       - name: Test with Coverage
         run: pdm run pytest tests --cov=app --cov-report=xml
+
+      - if: runner.os == 'Linux'
+        name: Stop Redis & Database Services
+        run: sudo docker compose -f tests/docker-compose.yml down --remove-orphans --volumes
+
 
       - uses: actions/upload-artifact@v4
         with:
@@ -75,12 +91,12 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.11"
+          python-version: "3.12"
 
       - uses: pdm-project/setup-pdm@v4
         name: Set up PDM
         with:
-          python-version: "3.11"
+          python-version: "3.12"
           allow-python-prereleases: true
           cache: true
 
@@ -120,7 +136,7 @@ jobs:
       - uses: pdm-project/setup-pdm@v4
         name: Set up PDM
         with:
-          python-version: "3.11"
+          python-version: "3.12"
           cache: true
       - name: Install dependencies
         run: pdm install -G:all
