+++ tests/docker-compose.yml
@@ -2,7 +2,7 @@ version: "3"
 
 services:
   postgres:
-    image: postgres
+    image: postgres:16
     ports:
       - "5423:5432"
     environment:
